feat: add new `custom` option for `target` property
Add a new `custom` option for the `target` property that allows users
to run a custom script given the `MATCHED_FILES` array. This enables
users to build more end-to-end pipelines, since the scope at which the
`custom` script can operate is not limited to each of the matched files,
like a `command` script is, but extends to the entire pipeline itself.